Moving ContinuationTokenItem to the types file
