+++ src/capturer/cosmos/cosmos.ts
@@ -8,8 +8,8 @@ import * as E from "fp-ts/Either";
 import * as TE from "fp-ts/TaskEither";
 import * as B from "fp-ts/boolean";
 import { pipe } from "fp-ts/lib/function";
-import { ProcessResult } from "../../factory/types";
-import { ContinuationTokenItem, upsertItem } from "./utils";
+import { ContinuationTokenItem, ProcessResult } from "../../factory/types";
+import { upsertItem } from "./utils";
 
 /**
  * Returns the options for creating a change feed iterator.
