+++ src/factory/service.ts
@@ -2,8 +2,7 @@ import { Container, CosmosClient, Database } from "@azure/cosmos";
 import * as O from "fp-ts/Option";
 import { TaskEither } from "fp-ts/lib/TaskEither";
 import { Collection, Db, Document, MongoClient } from "mongodb";
-import { ContinuationTokenItem } from "../capturer/cosmos/utils";
-import { ProcessResult } from "./types";
+import { ContinuationTokenItem, ProcessResult } from "./types";
 export interface IDatabaseConfig {
   readonly connection: string;
 }
